src: remove global symbol_table
Store symbol tables in context object instead. Use the nft_ctx object to
store the dynamic symbol table. Pass it on to the parse_ctx object so
this can be accessed from the parse routines. This dynamic symbol table
is also accesible from the output_ctx object for print routines.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>